add transaction_xlayer.go
